Rewrite fairly broken recursive fibonacci test to use @join_app
The previous implementation of this test could not recurse (for n>3)
when executed with a non-local executor such as htex, because
it relied on being able to launch apps from within arbitary
apps. In the new implementation, this is no longer the case
because @join_apps always run in the submitting process, not
on arbitrary executors.

The previous implementation of this test hangs when run on a
local executor when run with n sufficiently large to block all
of the threads in the local executor. In the new implementation,
that blocking is replaced by tasks being in th the 'joining' state,
releasing the local executor thread.